% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyChanges.R
\name{calcAndClassifyChanges}
\alias{calcAndClassifyChanges}
\title{Classify Changes}
\usage{
calcAndClassifyChanges(
  data_dt,
  comparisons_lsv,
  idCol_v = "Time point",
  metaCols_v = c("RECIST", "CB"),
  mergeCol_v = "Patients",
  measureName_v = "population",
  change_v = 10
)
}
\arguments{
\item{data_dt}{data in melted format (1 row per patient + timepoint + population)}

\item{comparisons_lsv}{list of comparisons to calculate. List name is name of comparison, values are columns to compare}

\item{idCol_v}{column that identifies unique measurements (time for us usually)}

\item{metaCols_v}{meta.data columns to ignore during merging (they'll be the same in all tables)}

\item{measureName_v}{name to give the variable column (tbd d)}

\item{change_v}{percentage change to use for bins}

\item{mergecol_v}{column to merge tables on}
}
\value{
data.table
}
\description{
classify changes between cycles for various populations
}
\details{
There might be an easier way to do this with fewer lines, but haven't looked into it.
For all of the comparisons in comparison_lsv, calculate the difference between the two
time columns, bin it, and then combine them all together
}
