% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbColors_v}
\alias{cbColors_v}
\title{Clinical Benefit Colors}
\format{
\subsection{'cbColors_v'}{

Named vector of length 2
\describe {
\item{values}{Hex Color}
\item{names}{Clinical Benefit Group}
}
}
}
\usage{
cbColors_v
}
\description{
Colors to differentiate Clinical Benefit responses
}
\keyword{datasets}
