% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbTimeSeries.R
\name{cbTimeSeries}
\alias{cbTimeSeries}
\title{Pairwise Boxplot}
\usage{
cbTimeSeries(
  data_dt,
  yVar_v,
  xVar_v = "Time point",
  pairVar_v = "Patients",
  shapeVar_v = "Patients",
  lineVar_v = "CB",
  colorVar_v = "CB",
  colorColors_v = cbColors_v,
  fillVar_v = "Time point",
  fillColors_v = timeColors_v
)
}
\arguments{
\item{data_dt}{table of population values and a few meta columns}

\item{yVar_v}{variable to use for y-axis (should be some population column name, e.g. "Live cells")}

\item{xVar_v}{variable to put on x-axis (usually 'Time point')}

\item{pairVar_v}{variable to connect paired samples (usually "Patients")}

\item{shapeVar_v}{variable to shape points by (usually "Patients")}

\item{lineVar_v}{variable to change linetypes}

\item{colorVar_v}{variable to color LINES by (usually "Patients")}

\item{colorColors_v}{named color vector to use as color scale for color variable}

\item{fillVar_v}{variable to color BOXES by (usually "Time point")}

\item{fillColors_v}{named color vector to use as color scale for fill variable}
}
\value{
ggplot
}
\description{
Pairwise boxplot for clinical data
}
