% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cd48Markers_lsv}
\alias{cd48Markers_lsv}
\title{CD4/CD8 Markers List}
\format{
\subsection{'cd48Markers_lsv'}{

List of length 3
\describe {
\item{names}{placeholder separater}
\item{values}{Marker}
}
}
}
\usage{
cd48Markers_lsv
}
\description{
Markers of CD4+ and CD8+ Cells (split into groups)
}
\keyword{datasets}
