% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cd48Markers_v}
\alias{cd48Markers_v}
\title{CD4/CD8 Markers}
\format{
\subsection{'cd48Markers_v'}{

Vector of length 9
\describe{
\item{values}{Marker}
}
}
}
\usage{
cd48Markers_v
}
\description{
Markers of CD4+ and CD8+ Cells
}
\keyword{datasets}
