% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{cleanData}
\alias{cleanData}
\title{Clean Data}
\usage{
cleanData(data_dt, metaCols_v = NULL)
}
\arguments{
\item{data_dt}{input table exported from FloJo}

\item{metaCols_v}{optional columns that need to not be converted to numeric}
}
\value{
input table with the above changes
}
\description{
Clean up data from the FloJo output
}
\details{
Clean up data by
\enumerate{
\item Removing 'Mean' and 'SD' rows
\item Replacing 'n/a' and 'n/a \%' with actual NA values
\item Change percentages to actual values
\item Remove empty columns
}
}
