% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{cleanFuncCols}
\alias{cleanFuncCols}
\title{Clean Functional Columns}
\usage{
cleanFuncCols(data_dt, metaCols_v = NULL)
}
\arguments{
\item{data_dt}{input table exported from FloJo}

\item{metaCols_v}{optional metadata columns that don't need fixing}
}
\value{
input table with new column names
}
\description{
Simplify the names for each population(s) functional markers from the FloJo output
}
\details{
Clean up columns by
\enumerate{
\item removing everything after the |
\item Split on slash and grab last two to get population and func marker
\item Remove negative functional markers (i.e. only keep CD38+, not CD38-)
\item Get rid of trailing +'s for easier grep
e.g. Debris exclusion/Single Cells/Live cells/CD45+/Non-B cells/ab CD3+ T cells/CD8+ T cells/Activated CD8+ CD45RO+ T cells/CD8+ Tcm/CD39+ | Freq. of Parent
turns into CD39 in the CD8 Tcm table
}
}
