% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{cleanGmfiCols}
\alias{cleanGmfiCols}
\title{Clean GMFI Columns}
\usage{
cleanGmfiCols(data_dt)
}
\arguments{
\item{data_dt}{input table exported from FloJo}
}
\value{
input table with new column names
}
\description{
Simplify the names for each population and marker from the FloJo output
}
\details{
Clean up columns by
\enumerate{
\item Extract population - after last / and before |
\item Get the marker in the parenthesis at the end: (marker)
\item Make new column combining that info
}
}
