% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{cleanStackedCols}
\alias{cleanStackedCols}
\title{Clean Stacked Columns}
\usage{
cleanStackedCols(data_dt)
}
\arguments{
\item{data_dt}{input table exported from FloJo}
}
\value{
input table with new column names
}
\description{
Simplify the names for each population and marker from the FloJo output
}
\details{
Clean up columns by
\enumerate{
\item Extract population - after last / and before |
}
}
