% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors_v}
\alias{colors_v}
\title{Colors}
\format{
\subsection{'colors_v'}{

Named vector of length 56
\describe {
\item{values}{Hex Code}
\item{names}{population/marker name}
}
}
}
\usage{
colors_v
}
\description{
Population and Marker colors
}
\keyword{datasets}
