% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compactSchema_dt}
\alias{compactSchema_dt}
\title{Compact Schema}
\format{
\subsection{'compactSchema_dt'}{

data.table identifying treatments given on each day
\describe {
\item{A}{Day}
\item{B}{Treatment(s) given that day}
}
}
}
\usage{
compactSchema_dt
}
\description{
One-row table of treatment schema
}
\keyword{datasets}
