% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSchemaTableGrob.R
\name{createDiagonalCellGrob}
\alias{createDiagonalCellGrob}
\title{Create Diagonal Cell Grob}
\usage{
createDiagonalCellGrob(
  x_v,
  y_v,
  width_v,
  height_v,
  direction_v = "forward",
  topColor_v,
  bottomColor_v
)
}
\arguments{
\item{x_v}{row value of cell that needs a diagonal}

\item{y_v}{column value of cell that needs a diagonal}

\item{width_v}{width of cell}

\item{height_v}{height of cell}

\item{direction_v}{Direction of the triangle. Either 'forward' for a forward slash (/) or 'backward" for backslash}

\item{topColor_v}{color vector for top portion}

\item{bottomColor_v}{color vector for bottom portion}
}
\value{
tbd
}
\description{
Make a diagonal cell inside of a gtable
}
\details{
tbd
}
