% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSchemaTableGrob.R
\name{createSchemaTableGrob}
\alias{createSchemaTableGrob}
\title{Create Schema Table Grob}
\usage{
createSchemaTableGrob(data_dt, colorMap_dt, direction_v = "forward")
}
\arguments{
\item{data_dt}{input table to be displayed}

\item{colorMap_dt}{reference table that has a color for each cell designation. If data_dt has value not in colorMap, it will not get filled}

\item{direction_v}{passed to drawDiagonalCellGrob to determine triangle direction (forward or backward for / or \ respectively)}
}
\value{
tbd
}
\description{
Make a gtable with specified cells colored appropriately
}
\details{
tbd
}
