% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycles_lsv}
\alias{cycles_lsv}
\title{Cycles and Days}
\format{
\subsection{'cycles_lsv'}{

Named list of length 4
\describe {
\item{values}{Cycle/Day designations}
\item{names}{Cycle designation}
}
}
}
\usage{
cycles_lsv
}
\description{
List of each cycle and the day measurements it contains
}
\keyword{datasets}
