% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycles_v}
\alias{cycles_v}
\title{Cycles}
\format{
\subsection{'cycles_v'}{

Vector of length 7
\describe {
\item{values}{Cycle}
}
}
}
\usage{
cycles_v
}
\description{
Base names of each cycle that have measurements
}
\keyword{datasets}
