% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dirColors_v}
\alias{dirColors_v}
\title{Dir Colors}
\format{
\subsection{'dirColors_v'}{

Named vector of length 3
\describe{
\item{values}{Hex Color}
\item{names}{Direction}
}
}
}
\usage{
dirColors_v
}
\description{
Colors to differentiate direction of measurement change
}
\keyword{datasets}
