% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSchemaTable.R
\name{drawDiagonalCell}
\alias{drawDiagonalCell}
\title{Draw Diagonal Cell}
\usage{
drawDiagonalCell(
  x_v,
  y_v,
  width_v,
  height_v,
  topColor_v,
  bottomColor_v,
  direction_v = "backward"
)
}
\arguments{
\item{x_v}{row value of cell that needs a diagonal}

\item{y_v}{column value of cell that needs a diagonal}

\item{width_v}{width of cell}

\item{height_v}{height of cell}

\item{topColor_v}{color vector for top portion}

\item{bottomColor_v}{color vector for bottom portion}

\item{direction_v}{Direction of the triangle. Either 'forward' for a forward slash (/) or 'backward" for backslash}
}
\value{
tbd
}
\description{
Make a diagonal cell inside of a gtable
}
\details{
tbd
}
