% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fullSchema_dt}
\alias{fullSchema_dt}
\title{Full Schema}
\format{
\subsection{'fullSchema_dt'}{

data.table identifying treatments given on each day
\describe{
\item{A}{Day}
\item{B}{Axatilimab (aCSF1R) treatments}
\item{C}{Retifanlimab (aPD1) treatments}
\item{D}{Paclitaxel (PTX) treatments}
}
}
}
\usage{
fullSchema_dt
}
\description{
Multi-row table of treatment schema
}
\keyword{datasets}
