% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalLinePlot.R
\name{functionalLine}
\alias{functionalLine}
\title{Prop and Func Bar}
\usage{
functionalLine(
  func_dt,
  markerCol_v = "marker",
  markers_v,
  markerColors_v,
  patientCol_v = "Patients",
  pt_v,
  classCol_v = "variable",
  xVar_v = "Time point",
  yVar_v = "value",
  title_v
)
}
\arguments{
\item{func_dt}{functional table}

\item{markerCol_v}{column for functional markers (should always be 'marker')}

\item{markers_v}{which markers to plot}

\item{markerColors_v}{color vector for marker color scale}

\item{patientCol_v}{column for patients (should always be 'Patients')}

\item{pt_v}{which patient to plot}

\item{classCol_v}{column that holds cell classes. Used for facet}

\item{xVar_v}{column to put on x-axis. Should be 'Time point'}

\item{yVar_v}{column to use for y-axis. Should be 'value'}

\item{title_v}{title for plot}
}
\description{
Line/spaghetti plot of functional expression
}
