% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indTimeColors_lsv}
\alias{indTimeColors_lsv}
\title{Individual Time Colors}
\format{
\subsection{'indTimeColors_lsv'}{

Named list of length 5
\describe{
\item{values}{Color-blind friendly colors}
\item{names}{name of cycle or day set}
}
}
}
\usage{
indTimeColors_lsv
}
\description{
Colors for each cycle for specific plots
}
\keyword{datasets}
