% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotRadar.R
\name{makeRadar}
\alias{makeRadar}
\title{Build one radar chart}
\usage{
makeRadar(
  curr_dt,
  currTime_v,
  patientCol_v,
  patientColors_v,
  legend_v,
  maxPt_v,
  globalMin_v,
  globalMid_v,
  globalMax_v
)
}
\arguments{
\item{curr_dt}{data.table with the data for one time point}

\item{currTime_v}{character, title of the plot}

\item{patientCol_v}{column that identifies the patient / group}

\item{patientColors_v}{named vector with colours}

\item{legend_v}{logical indicating if legend should be printed.}

\item{maxPt_v}{tbd}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{          ggplot object
}\if{html}{\out{</div>}}
}
\description{
Build one radar chart
}
