% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientTimeSeries.R
\name{patientTimeSeries}
\alias{patientTimeSeries}
\title{Pairwise Boxplot}
\usage{
patientTimeSeries(
  data_dt,
  yVar_v,
  xVar_v = "Time point",
  D1Lines_v = T,
  pairVar_v = "Patients",
  shapeVar_v = "Patients",
  lineVar_v = "Patients",
  colorVar_v = "Patients",
  colorColors_v = ptColors_v,
  fillVar_v = "Time point",
  fillColors_v = timeColors_v
)
}
\arguments{
\item{data_dt}{table of population values and a few meta columns}

\item{yVar_v}{variable to use for y-axis (should be some population column name, e.g. "Live cells")}

\item{xVar_v}{variable to put on x-axis (usually 'Time point')}

\item{D1Lines_v}{logical indicating if vertical grey lines should be made at each D1 (to help distinguish cycle starts)}

\item{pairVar_v}{variable to connect paired samples (usually "Patients")}

\item{shapeVar_v}{variable to shape points by (usually "Patients")}

\item{lineVar_v}{variable to change linetypes (usually "Patients")}

\item{colorVar_v}{variable to color LINES by (usually "Patients")}

\item{colorColors_v}{named color vector to use as color scale for color variable}

\item{fillVar_v}{variable to color BOXES by (usually "Time point")}

\item{fillColors_v}{named color vector to use as color scale for fill variable}
}
\value{
list of ggplots
}
\description{
Pairwise boxplot for clinical data
}
