% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCycleCompareLine.R
\name{plotCycleCompareLine}
\alias{plotCycleCompareLine}
\title{Plot Cycle Compare Line}
\usage{
plotCycleCompareLine(
  data_dt,
  dataCols_v,
  colors_v = c(UP = "#47AD48", DN = "#D7222B", NC = "#0A0B09"),
  xVar_v = "variable",
  yVar_v = "value",
  lineVar_v = "CB",
  groupVar_v = "Patients",
  colorVar_v = "Direction"
)
}
\arguments{
\item{data_dt}{data.table that has been run through calcAndClassifyChanges (must have Compare column)}

\item{dataCols_v}{columns to plot}

\item{colors_v}{color vector for direction}

\item{xVar_v}{x-axis variable (should always be 'variable' if calcAndClassifyChanges has been run)}

\item{yVar_v}{y-axis variable (should always be 'value' if calcAndClassifyChanges has been run)}

\item{lineVar_v}{variable to modify linetype. Should be "CB" or "RECIST" likely}

\item{groupVar_v}{variable to connect lines}

\item{colorVar_v}{name of variable to color by (should be 'Direction' if calcAndClassifyChanges has been run)}
}
\value{
list of plots (one for each element in dataCols_v)
}
\description{
Pairwise lineplot between different timepoints,
with lines colored by their change direction
}
\details{
This should only be run after running calcAndClassifyChanges, otherwise it's doubtful that data_dt will be
correct
}
