% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popLinePlot.R
\name{popLine}
\alias{popLine}
\title{Prop and Func Bar}
\usage{
popLine(
  bar_dt,
  popCol_v = "variable",
  pop_v,
  patientCol_v = "Patients",
  ptColors_v,
  xVar_v = "Time point",
  yVar_v = "value",
  yLab_v = "Pct CD45",
  title_v
)
}
\arguments{
\item{bar_dt}{barchart table (melted)}

\item{popCol_v}{column that holds populations. Should be 'variable' in melted data}

\item{pop_v}{which population to plot}

\item{patientCol_v}{column for patients (should always be 'Patients')}

\item{ptColors_v}{named color vector for patients}

\item{xVar_v}{column to put on x-axis. Should be 'Time point'}

\item{yVar_v}{column to use for y-axis. Should be 'value'}

\item{title_v}{title for plot}

\item{markerColors_v}{color vector for marker color scale}
}
\value{
ggplot
}
\description{
Line/spaghetti plot of functional expression
}
