% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propAndFuncBar.R
\name{propAndFuncBar}
\alias{propAndFuncBar}
\title{Prop and Func Bar}
\usage{
propAndFuncBar(
  data_dt,
  func_dt,
  markerCol_v = "marker",
  markers_v,
  markerColors_v,
  patientCol_v = "Patients",
  pt_v,
  timeCol_v = "Time point",
  time_v,
  xVar_v = "variable",
  yVar_v = "value",
  title_v,
  allOut_v = F
)
}
\arguments{
\item{data_dt}{cell proportion table}

\item{func_dt}{functional table}

\item{markerCol_v}{column for functional markers (should always be 'marker')}

\item{markers_v}{which markers to plot}

\item{markerColors_v}{color vector for marker color scale}

\item{patientCol_v}{column for patients (should always be 'Patients')}

\item{pt_v}{which patient to plot}

\item{timeCol_v}{column for timepoints (should always be 'Time point')}

\item{time_v}{which time to plot}

\item{title_v}{title for plot}

\item{allOut_v}{logical indicating whether to output list of stuff or just the combined plot}
}
\value{
either list or ggplot
}
\description{
Combo plot showing stacked bar of cell proportions and
regular bar of marker functionalitty
}
