% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptColors_v}
\alias{ptColors_v}
\title{Patient Colors}
\format{
\subsection{'ptColors_v'}{

Named vector of length 5
\describe{
\item{values}{Hex Color}
\item{names}{Patient ID}
}
}
}
\usage{
ptColors_v
}
\description{
Colors to differentiate patients
}
\keyword{datasets}
