% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{radarMarkers_v}
\alias{radarMarkers_v}
\title{Radar Markers}
\format{
\subsection{'radarMarkers_v'}{

Vector of length 7
\describe {
\item{values}{Marker}
}
}
}
\usage{
radarMarkers_v
}
\description{
Markers for radar plots
}
\keyword{datasets}
