% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recistColors_v}
\alias{recistColors_v}
\title{RECIST Colors}
\format{
\subsection{'recistColors_v'}{

Named vector of length 3
\describe {
\item{values}{Hex Color}
\item{names}{RECIST Group}
}
}
}
\usage{
recistColors_v
}
\description{
Colors to differentiate RECIST responses
}
\keyword{datasets}
