% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schemaColorMap_dt}
\alias{schemaColorMap_dt}
\title{Color Map Table}
\format{
\subsection{'compactSchema_dt'}{

data.table identifying treatments given on each day
\describe {
\item{Tx}{Treatment (or treatment combination)}
\item{Color}{Hex code}
}
}
}
\usage{
schemaColorMap_dt
}
\description{
data.table to map color names (including combo names) with hex codes
}
\keyword{datasets}
