% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schemaColorMap_lsv}
\alias{schemaColorMap_lsv}
\title{Color Map List}
\format{
\subsection{'schemaColorMap_lsv'}{

data.table identifying treatments given on each day
\describe{
\item{names}{Treatment (or treatment combination)}
\item{values}{Hex code}
}
}
}
\usage{
schemaColorMap_lsv
}
\description{
list to map color names (including combo names) with hex codes
Same as schemaColorMap_dt, but list form
}
\keyword{datasets}
