% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schemaColorMap_v}
\alias{schemaColorMap_v}
\title{Color Map Vector}
\format{
\subsection{'schemaColorMap_v'}{

vector with treatment colors
\describe{
\item{names}{Treatment (or treatment combination)}
\item{values}{Hex code}
}
}
}
\usage{
schemaColorMap_v
}
\description{
vector to map color names (including combo names) with hex codes
Same as schemaColorMap_dt, but vector form
}
\keyword{datasets}
