% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sheetRename_dt}
\alias{sheetRename_dt}
\title{Sheet Rename}
\format{
\subsection{'sheetRename_dt'}{

data.table mapping short and long names
\describe{
\item{Short}{Shortened name}
\itme{Long}{Original (long) name}
}
}
}
\usage{
sheetRename_dt
}
\description{
Map table to rename populations that get cut off by excel
}
\keyword{datasets}
