% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shortNameColors_v2_v}
\alias{shortNameColors_v2_v}
\title{Short Colors2}
\format{
\subsection{'shortNameColors_v2_v'}{

Named vector of length 56
\describe {
\item{values}{Hex Code}
\item{names}{shortened population/marker name}
}
}
}
\usage{
shortNameColors_v2_v
}
\description{
Version2 of Short Population and Marker colors
}
\keyword{datasets}
