% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortTimePoints.R
\name{sortTimePoints}
\alias{sortTimePoints}
\title{Sort Time Points}
\usage{
sortTimePoints(timePoints_v)
}
\arguments{
\item{timePoints_v}{vector of time points used}
}
\value{
ordered vector
}
\description{
Sort timepoints into correct order
}
\details{
this is specifically for timepoints of C#D# format
(e.g. C1D1, C2D8, etc.). First sort by cycle, then by days within each cycle
}
