% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitMinorPops}
\alias{splitMinorPops}
\title{Split Minor Pops}
\usage{
splitMinorPops(
  data_dt,
  metaCols_v = NULL,
  map_lsv = list(Bcells = c("CD19+ CD20+ B cells",
    "CD19+CD20- Plasmablasts Plasma cells"), Tcells = c("CD4+ T cells", "CD8+ T cells",
    "CD8+ CD4+ double positive T cells", "CD8- CD4- double negative T cells"), NKcells =
    c("CD56bright CD16- NK cells", "CD56dim CD16+ NK cells", "CD56dim CD16- NK cells"),
    MyeloidAPCs = c("CD14+ CD16+ intermediary Monocytes",
    "CD14+ CD16- classical Monocytes", "CD14dim CD16+ non-classical Monocytes",
    "CD11c+ Myeloid DCs", "CD123+ CD11c+ plasmacytoid DCs"), MyeloidDCs = c("CD1c+ cDC2",
    
     "CD141+ cDC1"), Granulocytes = c("CD15+ CD49d+ eosinophils",
    "CD15+ CD123+ basophils", "CD15+ putative neutrophils"))
)
}
\arguments{
\item{data_dt}{raw data}

\item{metaCols_v}{optional metadata columns to keep for each}

\item{map_lsv}{named list indicating which populations to put into which output data.table}
}
\value{
list of data.tables made up of input data_dt
}
\description{
Split into different tables
}
