% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitTSubsets}
\alias{splitTSubsets}
\title{Split T Subsets}
\usage{
splitTSubsets(data_dt, split_v = c("CD4", "CD8"))
}
\arguments{
\item{data_dt}{raw data}
}
\value{
list of data.tables made up of input data_dt
}
\description{
Split into different tables
}
