% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeColors_v}
\alias{timeColors_v}
\title{Time Colors}
\format{
\subsection{'timeColors_v'}{

Named vector of length 17
\describe {
\item{values}{Hex Color}
\item{names}{Timepoint}
}
}
}
\usage{
timeColors_v
}
\description{
Colors to differentiate timepoints
}
\keyword{datasets}
