% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timePoints_v}
\alias{timePoints_v}
\title{Timepoints}
\format{
\subsection{'timePoints_v'}{

Vector of length 17
\describe{
\item{values}{Timepoint}
}
}
}
\usage{
timePoints_v
}
\description{
Timepoints of patient measuremenets
}
\keyword{datasets}
