% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatDateMap_dt}
\alias{treatDateMap_dt}
\title{Treatment-Date Map Table}
\format{
\subsection{'treatDateMap_dt'}{

data.table identifying treatments given on each day/cycle and their color
\describe{
\item{Time points}{CycleDay time points}
\item{Cycle}{treatment cycle}
\item{Day}{treatment day (within cycle)}
\item{Tx}{Treatment (or treatment combination)}
\item{Color}{Hex code}
}
}
}
\usage{
treatDateMap_dt
}
\description{
data.table to map time points (date) with cycle, day, treatment, and color
}
\keyword{datasets}
