% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotRadar.R
\name{ggPlotRadar}
\alias{ggPlotRadar}
\title{Plot Radar}
\usage{
ggPlotRadar(
  data_dt,
  markerCol_v = "marker",
  markers_v,
  patientCol_v = "Patients",
  patientColors_v,
  timeCol_v = "Time point",
  time_v = NULL,
  name_v = NULL,
  file_v = NULL
)
}
\arguments{
\item{data_dt}{data.table with functional markers}

\item{markerCol_v}{column in data_dt that identifies markers that can be plotted (should be 'marker')}

\item{markers_v}{character vector indicating which markers in markerCol_v to include in the plot}

\item{patientCol_v}{which column identifies different patient observations (should be 'Patients')}

\item{patientColors_v}{color vector for the patients}

\item{timeCol_v}{which column in data_dt identifies different time points. One radar plot per time}

\item{time_v}{optional vector to only plot specific time points instead of all available}

\item{name_v}{optional vector to label the entire plot (e.g. CD4+ Tcm)}

\item{file_v}{optional file to save output. If NULL, will print to console}
}
\value{
returns NULL (print to console). Save pdf if file_v is provided.
}
\description{
Wrapper for FSMB Radar plot
}
